fix precedence when default value is non-empty
The only case this actually applies to is the delimiter, which has a
default value of '-', however all others changed to make them
consistent.

This fixes the case where a module label is instantiated with a
non-default delimiter, and then another label is generated based off the
context but uses the default delimiter as the local var takes
precendance.